Fix volatile knowledge errors
