--- src/SerenityStar.SDK/Models/Connector/ConnectorStatusResponse.cs
@@ -1,9 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace SerenityStar.Models.Connector;
-
-public class ConnectorStatusResponse
-{
-    [JsonPropertyName("isConnected")]
-    public bool IsConnected { get; set; }
-}
