--- src/SerenityStar.SDK/Models/Connector/GetConnectorStatusOptions.cs
@@ -1,7 +1,0 @@
-namespace SerenityStar.Models.Connector;
-
-public class GetConnectorStatusOptions
-{
-    public Guid AgentInstanceId { get; set; }
-    public Guid ConnectorId { get; set; }
-}
