--- src/SerenityStar.SDK/Models/MessageFeedback/MessageFeedbackRequest.cs
@@ -1,9 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace SerenityStar.Models.MessageFeedback;
-
-public class MessageFeedbackRequest
-{
-    [JsonPropertyName("feedback")]
-    public bool Feedback { get; set; }
-}
