--- src/SerenityStar.SDK/Models/MessageFeedback/RemoveFeedbackOptions.cs
@@ -1,6 +1,0 @@
-namespace SerenityStar.Models.MessageFeedback;
-
-public class RemoveFeedbackOptions
-{
-    public Guid AgentMessageId { get; set; }
-}
