--- src/SerenityStar.SDK/Models/MessageFeedback/SubmitFeedbackOptions.cs
@@ -1,7 +1,0 @@
-namespace SerenityStar.Models.MessageFeedback;
-
-public class SubmitFeedbackOptions
-{
-    public Guid AgentMessageId { get; set; }
-    public bool Feedback { get; set; }
-}
