+++ src/SerenityStar/Agents/AgentsScope.cs
@@ -1,5 +1,6 @@
 using SerenityStar.Agents.Conversational;
 using SerenityStar.Agents.System;
+using SerenityStar.Agents.VolatileKnowledge;
 using System.Net.Http;
 
 namespace SerenityStar.Agents
@@ -29,12 +30,20 @@ namespace SerenityStar.Agents
         /// </summary>
         public AssistantsScope Assistants { get; }
 
+        /// <summary>
+        /// Manage volatile knowledge.
+        /// </summary>
+        public VolatileKnowledgeScope VolatileKnowledge { get; }
+
         internal AgentsScope(HttpClient httpClient)
         {
+            // Agents
             Activities = new ActivitiesScope(httpClient);
             ChatCompletions = new ChatCompletionsScope(httpClient);
             Proxies = new ProxiesScope(httpClient);
             Assistants = new AssistantsScope(httpClient);
+
+            VolatileKnowledge = new VolatileKnowledgeScope(httpClient);
         }
     }
 }
