+++ src/SerenityStar/Client/SerenityClient.cs
@@ -2,16 +2,8 @@ using Microsoft.Extensions.Options;
 using SerenityStar.Agents;
 using SerenityStar.Constants;
 using SerenityStar.Models;
-using SerenityStar.Models.VolatileKnowledge;
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Net.Http;
-using System.Net.Http.Json;
-using System.Text.Json;
-using System.Text.Json.Serialization;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace SerenityStar.Client
 {
@@ -19,11 +11,6 @@ namespace SerenityStar.Client
     public class SerenityClient : ISerenityClient
     {
         private readonly HttpClient _httpClient;
-        private static readonly JsonSerializerOptions JsonOptions = new JsonSerializerOptions
-        {
-            PropertyNameCaseInsensitive = true,
-            DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull
-        };
 
         /// <summary>
         /// Interact with the different agents available in Serenity Star.
@@ -79,103 +66,5 @@ namespace SerenityStar.Client
             client.BaseAddress = new Uri(baseUrl);
             client.Timeout = TimeSpan.FromSeconds(timeout ?? 100);
         }
-
-        /// <inheritdoc />
-        public async Task<VolatileKnowledge> UploadVolatileKnowledgeAsync( // sarasa esto vuela de acá y revisar la lectura de la response (copiar del hub)
-            UploadVolatileKnowledgeReq request,
-            bool processEmbeddings = true,
-            bool noExpiration = false,
-            int? expirationDays = null,
-            CancellationToken cancellationToken = default)
-        {
-            request.Validate();
-
-            using MultipartFormDataContent content = new();
-
-            // Add content if provided
-            if (!string.IsNullOrEmpty(request.Content))
-                content.Add(new StringContent(request.Content), "Content");
-
-            // Add file if provided
-            if (request.FileStream is not null)
-            {
-                StreamContent fileContent = new(request.FileStream);
-                string contentType = GetContentType(request.FileName);
-                fileContent.Headers.ContentType = new System.Net.Http.Headers.MediaTypeHeaderValue(contentType);
-                content.Add(fileContent, "File", request.FileName);
-            }
-
-            // Add callback URL if provided
-            if (!string.IsNullOrEmpty(request.CallbackUrl))
-                content.Add(new StringContent(request.CallbackUrl), "CallbackUrl");
-
-            // Build query parameters
-            List<string> queryParams = new List<string>
-        {
-            $"processEmbeddings={processEmbeddings}",
-            $"noExpiration={noExpiration}"
-        };
-
-            if (expirationDays.HasValue)
-                queryParams.Add($"expirationDays={expirationDays.Value}");
-
-            string queryString = string.Join("&", queryParams);
-            string endpoint = $"volatileknowledge?{queryString}";
-
-            HttpResponseMessage response = await _httpClient.PostAsync(
-                endpoint,
-                content,
-                cancellationToken);
-
-            if (!response.IsSuccessStatusCode)
-            {
-                string errorContent = await response.Content.ReadAsStringAsync();
-                throw new HttpRequestException($"Request failed with status code {response.StatusCode}: {errorContent}");
-            }
-
-            VolatileKnowledge? volatileKnowledge = await response.Content.ReadFromJsonAsync<VolatileKnowledge>(cancellationToken: cancellationToken);
-
-            return volatileKnowledge ?? throw new InvalidOperationException("Failed to deserialize volatile knowledge response");
-        }
-
-        /// <inheritdoc />
-        public async Task<VolatileKnowledge> GetVolatileKnowledgeStatusAsync( // sarasa esto vuela de acá y revisar la lectura de la response (copiar del hub)
-            Guid knowledgeId,
-            CancellationToken cancellationToken = default)
-        {
-            HttpResponseMessage response = await _httpClient.GetAsync(
-                $"volatileknowledge/{knowledgeId}",
-                cancellationToken);
-
-            if (!response.IsSuccessStatusCode)
-            {
-                string errorContent = await response.Content.ReadAsStringAsync();
-                throw new HttpRequestException($"Request failed with status code {response.StatusCode}: {errorContent}");
-            }
-
-            VolatileKnowledge? volatileKnowledge = await response.Content.ReadFromJsonAsync<VolatileKnowledge>(cancellationToken: cancellationToken);
-
-            return volatileKnowledge ?? throw new InvalidOperationException("Failed to deserialize volatile knowledge response");
-        }
-
-        #region Private Methods
-        private static string GetContentType(string fileName)
-        {
-            string extension = Path.GetExtension(fileName).ToLowerInvariant();
-            return extension switch
-            {
-                ".pdf" => "application/pdf",
-                ".doc" => "application/msword",
-                ".docx" => "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-                ".txt" => "text/plain",
-                ".csv" => "text/csv",
-                ".md" => "text/markdown",
-                ".jpg" => "image/jpg",
-                ".jpeg" => "image/jpeg",
-                ".png" => "image/png",
-                _ => throw new NotSupportedException($"File extension '{extension}' is not supported for upload."),
-            };
-        }
-        #endregion
     }
 }
