+++ src/SerenityStar/Models/Connector/GetConnectorStatusOptions.cs
@@ -1,0 +1,19 @@
+using System;
+
+namespace SerenityStar.Models.Connector
+{
+    /// <summary>
+    /// Options for getting connector status.
+    /// </summary>
+    public class GetConnectorStatusOptions
+    {
+        /// <summary>
+        /// Gets or sets the agent instance ID.
+        /// </summary>
+        public Guid AgentInstanceId { get; set; }
+        /// <summary>
+        /// Gets or sets the connector ID.
+        /// </summary>
+        public Guid ConnectorId { get; set; }
+    }
+}
