+++ src/SerenityStar/Models/MessageFeedback/RemoveFeedbackOptions.cs
@@ -1,4 +1,6 @@
-namespace SerenityStar.Models.Conversation
+using System;
+
+namespace SerenityStar.Models.MessageFeedback
 {
     /// <summary>
     /// Options for removing feedback from a message.
@@ -8,6 +10,6 @@ namespace SerenityStar.Models.Conversation
         /// <summary>
         /// The ID of the agent message to remove feedback from.
         /// </summary>
-        public string AgentMessageId { get; set; } = string.Empty;
+        public Guid AgentMessageId { get; set; } = Guid.Empty;
     }
 }
